Try 2, task 3

+++ 0x0A-argc_argv/3-mul.c
@@ -10,7 +10,7 @@
 
 int main(int argc, char **argv)
 {
-	int i, mul = 1;
+	long int i, mul = 1;
 
 	if (argc < 2)
 	{
@@ -22,6 +22,6 @@ int main(int argc, char **argv)
 		for (i = 1; i < argc ; i++)
 			mul = mul * atoi(argv[i]);
 	}
-	printf("%d\n", mul);
+	printf("%ld\n", mul);
 	return (0);
 }
